src/python3requirements.txt
pyserial
urx
